requirements.txt
paho-mqtt
mysqlclient

